+++ src/keys/secret.rs
@@ -1,0 +1,63 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+//
+// Copyright (c) DUSK NETWORK. All rights reserved.
+
+use crate::JubJubScalar;
+use dusk_bytes::{Error, HexDebug, Serializable};
+use rand_core::{CryptoRng, RngCore};
+
+#[cfg(feature = "canon")]
+use canonical::Canon;
+#[cfg(feature = "canon")]
+use canonical_derive::Canon;
+
+#[allow(non_snake_case)]
+#[cfg_attr(feature = "canon", derive(Canon))]
+#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, HexDebug)]
+/// Structure repesenting a secret key
+pub struct SecretKey(pub(crate) JubJubScalar);
+
+impl From<JubJubScalar> for SecretKey {
+    fn from(s: JubJubScalar) -> SecretKey {
+        SecretKey(s)
+    }
+}
+
+impl From<&JubJubScalar> for SecretKey {
+    fn from(s: &JubJubScalar) -> SecretKey {
+        SecretKey(*s)
+    }
+}
+
+impl AsRef<JubJubScalar> for SecretKey {
+    fn as_ref(&self) -> &JubJubScalar {
+        &self.0
+    }
+}
+
+impl SecretKey {
+    /// This will create a random [`SecretKey`] from a scalar
+    /// of the Field JubJubScalar.
+    pub fn random<T>(rand: &mut T) -> SecretKey
+    where
+        T: RngCore + CryptoRng,
+    {
+        let fr = JubJubScalar::random(rand);
+
+        SecretKey(fr)
+    }
+}
+
+impl Serializable<32> for SecretKey {
+    type Error = Error;
+
+    fn to_bytes(&self) -> [u8; 32] {
+        self.0.to_bytes()
+    }
+
+    fn from_bytes(bytes: &[u8; 32]) -> Result<Self, Error> {
+        Ok(Self(JubJubScalar::from_bytes(bytes)?))
+    }
+}