+++ CHANGELOG.md
@@ -7,10 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.4.1] - 26-11-20
+## [0.5.0] - 2021-01-28
+
+### Added
+
+- Add `PublicKey` and `SecretKey` (removed from `schnorr`)
+- Add `dusk_bytes::Serializable` trait to structure
+
+### Removed
+
+- Remove manual implementation of `to_bytes` and `from_bytes`
+- Remove `Error` enum
+- Remove `decode` function
+
 ### Changed
+
+- Bump `dusk-jubjub` to `v0.8`
+- Bump `poseidon252` to `v0.16.0`
+- Bump `canonical` to `v0.5`
+- Bump `canonical_derive` `v0.5`
+- Update CHANGELOG to ISO 8601
+
+## [0.4.1] - 2020-11-26
+
+### Changed
+
 - Use poseidon252 dependency.
 
-## [0.4.0] - 17-11-20
+## [0.4.0] - 2020-11-17
+
 ### Changed
+
 - No-Std compatibility.